+++ windirstat/Controls/FileSearchControl.cpp
@@ -19,7 +19,7 @@
 #include "ItemSearch.h"
 #include "FileTreeView.h"
 
-CFileSearchControl::CFileSearchControl() : CTreeListControl(COptions::SearchViewColumnOrder.Ptr(), COptions::SearchViewColumnWidths.Ptr(), LF_SEARCHLIST, false)
+CFileSearchControl::CFileSearchControl() : CTreeListControl(COptions::SearchViewColumnOrder.Ptr(), COptions::SearchViewColumnWidths.Ptr(), COptions::SearchViewColumnVisibility.Ptr(), LF_SEARCHLIST, false)
 {
     m_singleton = this;
 }