+++ windirstat/res/LangStrings.h
@@ -1,5 +1,4 @@
 ﻿#pragma once
-#include <string_view>
 constexpr std::wstring_view IDS_ABOUT_ABOUT = L"IDS_ABOUT_ABOUT";
 constexpr std::wstring_view IDS_ABOUT_ABOUT_TEXTss = L"IDS_ABOUT_ABOUT_TEXTss";
 constexpr std::wstring_view IDS_ABOUT_LICENSE = L"IDS_ABOUT_LICENSE";
@@ -93,6 +92,7 @@ constexpr std::wstring_view IDS_IDLEMESSAGE = L"IDS_IDLEMESSAGE";
 constexpr std::wstring_view IDS_ITEMSs_SELECTED_FILESs_FOLDERSs = L"IDS_ITEMSs_SELECTED_FILESs_FOLDERSs";
 constexpr std::wstring_view IDS_JUNCTIONS = L"IDS_JUNCTIONS";
 constexpr std::wstring_view IDS_LARGEST_FILES = L"IDS_LARGEST_FILES";
+constexpr std::wstring_view IDS_MENU_ABSOLUTE_PERCENTAGES = L"IDS_MENU_ABSOLUTE_PERCENTAGES";
 constexpr std::wstring_view IDS_MENU_CANCEL = L"IDS_MENU_CANCEL";
 constexpr std::wstring_view IDS_MENU_CHKDSK = L"IDS_MENU_CHKDSK";
 constexpr std::wstring_view IDS_MENU_CLEANUP = L"IDS_MENU_CLEANUP";